Refactored sidebar to a presentational component
