+++ client/src/App.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { BrowserRouter as Router, Route, Switch, Redirect, withRouter } from 'react-router-dom';
-import { ThemeProvider } from 'styled-components';
+import { Route, Switch, Redirect } from 'react-router-dom';
 
 import { setInitialAuth } from './redux/actions/auth';
 import * as s from './styles';
@@ -19,13 +18,11 @@ import NotFound from './components/NotFound';
 import Navbar from './components/Presentational/Navbar';
 import PartyPage from './components/PartyPage';
 import Settings from './components/Container/Settings';
-import Sidebar from './components/Sidebar';
+import Sidebar from './components/Presentational/Sidebar';
 import NewRestaurant from './components/NewRestaurant';
 import RequireNotAuth from './components/HOC/RequireNotAuth';
 import RequireAuth from './components/HOC/RequireAuth';
-import { theme } from './global-styles/variables';
-
-const SidebarWithRouter = withRouter(props => <Sidebar {...props} />);
+import { sidebar } from './config/conditionalPathnames';
 
 const AuthedPartyPage = RequireAuth(PartyPage);
 const AuthedLoginEmployee = RequireAuth(LoginEmployee);
@@ -44,38 +41,39 @@ class App extends Component {
   }
 
   render() {
+    const { modalIsOpen, role, location, history } = this.props;
     return (
-      <Router>
-        <ThemeProvider theme={theme}>
-          <s.Container>
-            <Navbar blur={this.props.modalIsOpen} />
-            <s.Main>
-              <SidebarWithRouter modalIsOpen={this.props.modalIsOpen} role={this.props.role} />
-              <Switch>
-                <Route path="/" component={Landing} exact />
-                <Route path="/logout" component={Logout} />
-                <Route path="/login" component={NotAuthedLogin} />
-                <Route path="/register" component={NotAuthedRegsiter} />
-                <Route path="/success" component={RegistrationSuccess} />
-                <Route path="/login-employee" component={AuthedLoginEmployee} />
-                <Route path="/new-restaurant" component={AuthedNewRestaurant} />
-                <Route path="/new-employee" component={AuthedCreateEmployee} />
-                <Route path="/tables" component={AuthedTablesPage} />
-                <Route path="/servers" component={AuthedServers} />
-                <Route
-                  path="/party/:id"
-                  render={props => (
-                    <AuthedPartyPage {...props} modalIsOpen={this.props.modalIsOpen} />
-                  )}
-                />
-                <Route path="/settings" component={AuthedSettings} />
-                <Route path="/404" component={NotFound} exact />
-                <Redirect to="/404" />
-              </Switch>
-            </s.Main>
-          </s.Container>
-        </ThemeProvider>
-      </Router>
+      <s.Container>
+        <Navbar blur={modalIsOpen} />
+        <s.Main>
+          <Sidebar
+            blur={modalIsOpen}
+            role={role}
+            visible={!sidebar.includes(location.pathname)}
+            pathname={location.pathname}
+            push={history.push}
+          />
+          <Switch>
+            <Route path="/" component={Landing} exact />
+            <Route path="/logout" component={Logout} />
+            <Route path="/login" component={NotAuthedLogin} />
+            <Route path="/register" component={NotAuthedRegsiter} />
+            <Route path="/success" component={RegistrationSuccess} />
+            <Route path="/login-employee" component={AuthedLoginEmployee} />
+            <Route path="/new-restaurant" component={AuthedNewRestaurant} />
+            <Route path="/new-employee" component={AuthedCreateEmployee} />
+            <Route path="/tables" component={AuthedTablesPage} />
+            <Route path="/servers" component={AuthedServers} />
+            <Route
+              path="/party/:id"
+              render={props => <AuthedPartyPage {...props} modalIsOpen={modalIsOpen} />}
+            />
+            <Route path="/settings" component={AuthedSettings} />
+            <Route path="/404" component={NotFound} exact />
+            <Redirect to="/404" />
+          </Switch>
+        </s.Main>
+      </s.Container>
     );
   }
 }
