+++ client/src/components/Presentational/Sidebar/styles.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { NavLink } from 'react-router-dom';
 
-import { modalBlur } from '../../global-styles/mixins';
+import { modalBlur } from '../../../global-styles/mixins';
 
 export const Sidebar = styled.nav`
   display: ${(props) => (props.visible ? 'flex' : 'none')};
