+++ client/src/index.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import thunk from 'redux-thunk';
+import { BrowserRouter as Router, withRouter } from 'react-router-dom';
+import { ThemeProvider } from 'styled-components';
 import { createStore, applyMiddleware, compose } from 'redux';
 import { Provider } from 'react-redux';
 
@@ -8,6 +10,7 @@ import { requireManager } from './redux/middleware/permissions';
 import { axiosAuth } from './redux/middleware/axios';
 import reducer from './redux/reducers';
 import App from './App';
+import { theme } from './global-styles/variables';
 
 import './index.css';
 
@@ -26,9 +29,15 @@ if (process.env.NODE_ENV === 'production') {
   );
 }
 
+const AppWithRouter = withRouter(props => <App {...props} />);
+
 ReactDOM.render(
   <Provider store={store}>
-    <App />
+    <Router>
+      <ThemeProvider theme={theme}>
+        <AppWithRouter />
+      </ThemeProvider>
+    </Router>
   </Provider>,
   document.getElementById('root')
 );
